refactor: Relocate the previous implementation of GST
